+++ App/Sagas/subscribe/index.js
@@ -1,9 +1,7 @@
 // @flow
-import type {TChannel, TLike} from '../../../types/';
-import {channelsActions, userActions, defaultUser, likedChannelsActions} from '../../../Redux/';
-import {channelsRef, likesRef} from '../ref';
-import {firebaseApp} from '../init';
-import {convertUserFromFirebaseToStore} from '../index';
+import type {TChannel, TLike} from '../../types/';
+import {channelsActions, userActions, defaultUser, likedChannelsActions} from '../../Redux/';
+import {firebaseApp, channelsRef, likesRef, convertUserFromFirebaseToStore} from '../../Services/';
 
 const likes = (store: any, uid: string) => {
   const {dispatch, getState} = store;
